+++ client/header/activity-panel/index.js
@@ -14,33 +14,24 @@ import {
 import { getAdminLink } from '@woocommerce/wc-admin-settings';
 import { H, Section } from '@woocommerce/components';
 import {
-	NOTES_STORE_NAME,
 	ONBOARDING_STORE_NAME,
 	OPTIONS_STORE_NAME,
-	QUERY_DEFAULTS,
 	useUser,
 } from '@woocommerce/data';
 import { getHistory, getNewPath } from '@woocommerce/navigation';
 import { recordEvent } from '@woocommerce/tracks';
-import { applyFilters } from '@wordpress/hooks';
 
 /**
  * Internal dependencies
  */
 import './style.scss';
-import { getUnreadNotes } from './unread-indicators';
+import { getUnreadNotes, getUnreadNotifications } from './unread-indicators';
 import { isWCAdmin } from '../../dashboard/utils';
 import { Tabs } from './tabs';
 import { SetupProgress } from './setup-progress';
 import { DisplayOptions } from './display-options';
 import { HighlightTooltip } from './highlight-tooltip';
 import { Panel } from './panel';
-import {
-	getLowStockCount,
-	getOrderStatuses,
-	getUnreadOrders,
-} from '../../homescreen/activity-panel/orders/utils';
-import { getUnapprovedReviews } from '../../homescreen/activity-panel/reviews/utils';
 
 const HelpPanel = lazy( () =>
 	import( /* webpackChunkName: "activity-panels-help" */ './panels/help' )
@@ -72,28 +63,9 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 		return trackData;
 	};
 
-	const getIncompleteTasksCount = ( tasks, dismissedTasks ) => {
-		if ( ! tasks ) {
-			return 0;
-		}
-		return tasks.filter(
-			( task ) =>
-				task.visible &&
-				! task.completed &&
-				! dismissedTasks.includes( task.key )
-		).length;
-	};
-
-	const ALERTS_QUERY = {
-		page: 1,
-		per_page: QUERY_DEFAULTS.pageSize,
-		type: 'error,update',
-		status: 'unactioned',
-	};
-
 	const {
-		notifications,
 		hasUnreadNotes,
+		unreadNotifications = [],
 		requestingTaskListOptions,
 		setupTaskListComplete,
 		setupTaskListHidden,
@@ -101,48 +73,9 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 		previewSiteBtnTrackData,
 	} = useSelect( ( select ) => {
 		const { getOption, isResolving } = select( OPTIONS_STORE_NAME );
-		const { getNotes } = select( NOTES_STORE_NAME );
-		const orderStatuses = getOrderStatuses( select );
-		const countUnreadOrders = getUnreadOrders( select, orderStatuses );
-		const countLowStockProducts = getLowStockCount( select );
-		const countUnapprovedReviews = getUnapprovedReviews( select );
-		const storeAlerts = getNotes( ALERTS_QUERY );
-		const thingsToDoNext = applyFilters(
-			'woocommerce_admin_onboarding_task_list',
-			[],
-			query
-		);
-		const dismissedTasks = getOption(
-			'woocommerce_task_list_dismissed_tasks'
-		);
-
-		const storeAlertsCount = storeAlerts.length || 0;
-
-		const notificationsData = [
-			{
-				name: 'thingsToDoNext',
-				count:
-					getIncompleteTasksCount( thingsToDoNext, dismissedTasks ) +
-					storeAlertsCount,
-				critical: storeAlertsCount,
-			},
-			{
-				name: 'ordersToProcess',
-				count: countUnreadOrders,
-			},
-			{
-				name: 'reviewsToModerate',
-				count: countUnapprovedReviews,
-			},
-			{
-				name: 'stockNotices',
-				count: countLowStockProducts,
-			},
-		];
-
 		return {
 			hasUnreadNotes: getUnreadNotes( select ),
-			notifications: notificationsData,
+			unreadNotifications: getUnreadNotifications( select, query ),
 			requestingTaskListOptions:
 				isResolving( 'getOption', [
 					'woocommerce_task_list_complete',
@@ -213,7 +146,7 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 			name: 'inbox',
 			title: __( 'Inbox', 'woocommerce-admin' ),
 			icon: <Icon icon={ inboxIcon } />,
-			unread: hasUnreadNotes,
+			unread: hasUnreadNotes || unreadNotifications.length > 0,
 			visible:
 				( isEmbedded || ! isHomescreen() ) && ! isPerformingSetupTask(),
 		};
@@ -291,7 +224,7 @@ export const ActivityPanel = ( { isEmbedded, query, userPreferencesData } ) => {
 
 		switch ( tab ) {
 			case 'inbox':
-				return <InboxPanel notifications={ notifications } />;
+				return <InboxPanel notifications={ unreadNotifications } />;
 			case 'help':
 				return <HelpPanel taskName={ task } />;
 			default:
